--- a/mask.go
+++ b/mask.go
@@ -42,9 +42,11 @@ func calculateScore(mat *matrix.Matrix) int {
 	return score1 + score2 + score3 + score4
 }
 
-// 第一条规则为一行（或列）中的每组五个或更多相同颜色的模块提供QR代码。
+// check each row one-by-one. If there are five consecutive modules of the same color,
+// add 3 to the penalty. If there are more modules of the same color after the first five,
+// add 1 for each additional module of the same color. Afterward, check each column one-by-one,
+// checking for the same condition. Add the horizontal and vertical total to obtain penalty score
 func rule1(mat *matrix.Matrix) int {
-
 	var (
 		score          int
 		rowCurState    matrix.State
@@ -97,7 +99,10 @@ func rule1(mat *matrix.Matrix) int {
 	return score
 }
 
-// 第二个规则给出了QR码对矩阵中相同颜色模块的每个2x2区域的惩罚。
+// rule2
+// look for areas of the same color that are at least 2x2 modules or larger.
+// The QR code specification says that for a solid-color block of size m × n,
+// the penalty score is 3 × (m - 1) × (n - 1).
 func rule2(mat *matrix.Matrix) int {
 	var (
 		score          int
@@ -119,7 +124,6 @@ func rule2(mat *matrix.Matrix) int {
 	return score
 }
 
-// 如果存在看起来类似于取景器模式的模式，则第三规则给QR码一个大的惩罚
 // dark-light-dark-dark-dark-light-dark
 // 1011101 0000 or 0000 1011101
 //func rule3_backup(mat *matrix.Matrix) (score int) {
@@ -159,6 +163,11 @@ func rule2(mat *matrix.Matrix) int {
 //}
 
 // rule3 calculate punishment score in rule3, find pattern in QR Code matrix.
+// Looks for patterns of dark-light-dark-dark-dark-light-dark that have four
+// light modules on either side. In other words, it looks for any of the
+// following two patterns: 1011101 0000 or 0000 1011101.
+//
+// Each time this pattern is found, add 40 to the penalty score.
 func rule3(mat *matrix.Matrix) (score int) {
 	var (
 		pattern1     = binaryToStateSlice("1011101 0000")
@@ -190,7 +199,16 @@ func rule3(mat *matrix.Matrix) (score int) {
 	return score
 }
 
-// 如果超过一半的模块是暗的或轻的，则第四规则给QR码一个惩罚，对较大的差异有较大的惩罚
+// rule4 is based on the ratio of light modules to dark modules:
+//
+// 1. Count the total number of modules in the matrix.
+// 2. Count how many dark modules there are in the matrix.
+// 3. Calculate the percent of modules in the matrix that are dark: (darkmodules / totalmodules) * 100
+// 4. Determine the previous and next multiple of five of this percent.
+// 5. Subtract 50 from each of these multiples of five and take the absolute value of the result.
+// 6. Divide each of these by five. For example, 10/5 = 2 and 5/5 = 1.
+// 7. Finally, take the smallest of the two numbers and multiply it by 10.
+//
 func rule4(mat *matrix.Matrix) int {
 	var (
 		totalCnt             = mat.Width() * mat.Height()