- feat: matrix.Iterate with COLUMN is preferred.
- docs(qrcode): add comment to qrcode masking evaluation conditions.
- styl(qrcode): split utilities functions