@yeqown 2021-12-22
current progress

```sh
$ benchstat new1.txt new3.txt 
name              old time/op  new time/op  delta
_NewQRCode_1KB-4  11.1ms ±26%   7.7ms ± 8%  -30.83%  (p=0.000 n=9+8)
```