@yeqown 2021-12-23
```sh
$ benchstat new1.txt new5.txt
name              old time/op  new time/op  delta
_NewQRCode_1KB-4  11.1ms ±26%   6.9ms ± 9%  -38.09%  (p=0.000 n=9+9)

```